CB-338 refactor: add custom hook for getting platform information
플랫폼 정보를 얻는 커스텀 훅 추가
- 기존의 전역 상태 삭제
- 필요시마다 커스텀 훅 호출
